requirements.txt
pycrypto
pycryptodome
requests
mysql-connector-python
Django